adding package.json file
