+++ constructors/node.js
@@ -1,0 +1,6 @@
+function Node(data = null, next = null) {
+  this.data = data;
+  this.next = next;
+}
+
+module.exports = Node;
