+++ tests/tests.js
@@ -1,0 +1,3 @@
+const LL = require("../constructors/linked-list");
+
+console.log(LL);
